fix: Uniformized the code
Now it only reads the content 1 time with the new function '"buildHandlersAndEndpoints" in main
